refactor: test 코드 package 구조 변경
