+++ nextstep-di/src/test/java/nextstep/di/factory/example/config/ExampleConfig.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.config;
 
 import nextstep.annotation.Bean;
 import nextstep.annotation.Configuration;
@@ -7,19 +7,14 @@ import org.apache.commons.dbcp2.BasicDataSource;
 import javax.sql.DataSource;
 
 @Configuration
-public class IntegrationConfig {
+public class ExampleConfig {
     @Bean
     public DataSource dataSource() {
         BasicDataSource ds = new BasicDataSource();
         ds.setDriverClassName("org.h2.Driver");
-        ds.setUrl("\"jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE\"");
+        ds.setUrl("jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE");
         ds.setUsername("sa");
         ds.setPassword("");
         return ds;
-    }
-
-    @Bean
-    public MyJdbcTemplate jdbcTemplate(DataSource dataSource) {
-        return new MyJdbcTemplate(dataSource);
     }
 }
