+++ nextstep-di/src/test/java/nextstep/di/factory/example/config/IntegrationConfig.java
@@ -1,20 +1,26 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.config;
 
 import nextstep.annotation.Bean;
 import nextstep.annotation.Configuration;
+import nextstep.di.factory.example.repository.MyJdbcTemplate;
 import org.apache.commons.dbcp2.BasicDataSource;
 
 import javax.sql.DataSource;
 
 @Configuration
-public class ExampleConfig {
+public class IntegrationConfig {
     @Bean
     public DataSource dataSource() {
         BasicDataSource ds = new BasicDataSource();
         ds.setDriverClassName("org.h2.Driver");
-        ds.setUrl("jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE");
+        ds.setUrl("\"jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE\"");
         ds.setUsername("sa");
         ds.setPassword("");
         return ds;
+    }
+
+    @Bean
+    public MyJdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new MyJdbcTemplate(dataSource);
     }
 }
