+++ nextstep-di/src/test/java/nextstep/di/factory/example/controller/QnaController.java
@@ -1,6 +1,7 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.controller;
 
 import nextstep.annotation.Inject;
+import nextstep.di.factory.example.service.MyQnaService;
 import nextstep.stereotype.Controller;
 
 @Controller
