+++ nextstep-di/src/test/java/nextstep/di/factory/example/service/MyQnaService.java
@@ -1,6 +1,8 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.service;
 
 import nextstep.annotation.Inject;
+import nextstep.di.factory.example.repository.QuestionRepository;
+import nextstep.di.factory.example.repository.UserRepository;
 import nextstep.stereotype.Service;
 
 @Service
